feat: add more images to the news items with IDs 2 and 3
fchore: Corrects the navigation links in the secondary menu to direct to the correct pages.

chore: Corrects the image path in news item with IDs 2 and 3.

chore: refactor the style sheet for the news page was not found.

chore: change the width of the photos that apper at the top of the Agenda, FindOutMore and BoardOfDirectors Page.

chore: Refactor the font-size of the responsive navbar